feat(locales): add zh_tw.yml file for Chinese (Traditional) translation support
A new file `zh_tw.yml` has been added to the `config/locales` directory. This file contains translations for Chinese (Traditional) language. It is automatically loaded by Rails for internationalization purposes.

The file includes comments explaining how to use the translations with `I18n.t` and `t` methods in Rails views. It also provides an example of how to escape certain keys that may conflict with the default I18n backend.

The `zh-hant` section of the file includes translations for specific keys related to donation dates and parishioners in the context of a baptism.

This change was made to improve the user experience for Chinese (Traditional) speaking users by providing localized translations for the application.