feat(db): add migration to change index on special_donations table
The migration adds a new migration file `20240108084913_change_spe_donation_index.rb` which changes the index on the `special_donations` table.

In the `up` method, the migration removes the index on the columns `event_id` and `home_number` from the `special_donations` table.

In the `down` method, the migration adds the index back on the columns `event_id` and `home_number` in the `special_donations` table with the `unique` constraint.